# Add a api_platform.metadata_cache parameter