| Q             | A
| ------------- | ---
| Bug fix?      | no
| New feature?  | yes
| BC breaks?    | no
| Deprecations? | no
| Tests pass?   | yes
| Fixed tickets | n/a
| License       | MIT
| Doc PR        | n/a

Follows https://github.com/api-platform/core/issues/1724#issuecomment-367701990

TODO: fix tests

It's significant on my computer:

**Before**

Behat: 2m18.08s (2.16Gb)
PHPUnit: Time: 21.17 seconds, Memory: 104.25MB

**After**

Behat: 0m42.23s (498.70Mb)
Time: 25.72 seconds, Memory: 102.25MB

Can you try with your code base @jdeniau?