@jdeniau 2018-02-22
The tests finished to run on our server ~~and I'm a little disappointed by the results~~
For the record I ran the tests on our CI server, which is very busy at night with database replication and stuff like that. I will re-run the test, tomorrow or monday to give you "real day-to-day" data 

(I obfuscated the tests name for confidentiality reason)
## Without optimization: 
```
- 69 shuffled test classes into the queue.
- Will be consumed by 8 parallel Processes.
1	1/69	✔	1 s 250 ms	features/v.feature
7	2/69	✔	4 s 288 ms	features/w.feature
1	3/69	✔	2 min 27 s 196 ms	features/c.feature
7	4/69	✔	3 min 37 s 344 ms	features/o.feature
8	5/69	✔	4 min 18 s 4 ms	features/t.feature
6	6/69	✔	8 min 23 s 357 ms	features/p.feature
```
And then jenkins crashed 😉 

## With the `metadata_cache` set to true 
(I ran only the 6 tests)
```
- 6 shuffled test classes into the queue.
- Will be consumed by 8 parallel Processes.
3	1/6	✔	1 s 232 ms	features/v.feature
2	2/6	✔	2 s 920 ms	features/w.feature
1	3/6	✔	1 min 9 s 706 ms	features/c.feature
5	4/6	✔	1 min 46 s 215 ms	features/o.feature
4	5/6	✔	2 min 2 s 674 ms	features/t.feature
6	6/6	✔	3 min 34 s 540 ms	features/p.feature
```
The win is about 50%, which is not bad

## With kernel.debug set to false
```
- 6 shuffled test classes into the queue.
- Will be consumed by 8 parallel Processes.
6	1/6	✔	1 s 66 ms	features/v.feature
3	2/6	✔	2 s 783 ms	features/w.feature
1	3/6	✔	1 min 14 s 787 ms	features/c.feature
5	4/6	✔	1 min 52 s 721 ms	features/o.feature
4	5/6	✔	2 min 10 s 269 ms	features/t.feature
2	6/6	✔	3 min 47 s 400 ms	features/p.feature
```

Nearly the same time that with metadata_cache set to true

For the record, the "p.feature" file usually take 40 seconds to run, that's why I was disappointed ;)

I ran the `o.feature` test only on ouy development server and it took ~ 15 sec / 500 Mb with kernel.debug false or metadata_cache true and ~ 85 sec /  1.5 Gb without optimizations so big 👍 no matter what on the config point